backend-tally-dashboard/requirements.txt
# Django Core
Django==5.2
psycopg2-binary==2.9.11
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==24.2

# Configuration Management
python-decouple==3.8
python-dotenv==1.0.0

# Data Processing
pandas==2.3.3
numpy==2.3.4
openpyxl==3.1.5
et-xmlfile==2.0.0

# Celery & Task Queue
celery==5.5.3
redis==6.4.0
kombu==5.5.4
amqp==5.3.1
billiard==4.2.2
vine==5.1.0

# Image Processing
Pillow==12.0.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
click==8.3.0
click-didyoumean==0.3.1
click-repl==0.3.0
click-plugins==1.1.1.2
prompt-toolkit==3.0.52

# Authentication & Security
PyJWT==2.10.1
bcrypt==4.3.0
passlib==1.7.4

# Testing
pytest==8.0.2
pytest-django==4.8.0
Faker==37.8.0
factory-boy==3.3.0

# Development Tools
autopep8==2.3.2

# Other Dependencies
packaging==25.0
six==1.17.0
wcwidth==0.2.14
asgiref==3.10.0
sqlparse==0.5.3
